Refactor frame columns, add tab mock
