+++ ctui/component/contact_frame.py
@@ -1,0 +1,8 @@
+import urwid
+
+
+class CFrame(urwid.Frame):
+    def __init__(self, body, footer, core, config):
+        super(CFrame, self).__init__(body=body, footer=footer)
+        self.core = core
+        self.name = 'frame'
